Move FeedStore to its own file
